pyproject.toml
[tool.poetry]
name = "cosy"
version = "0.5.0"
description = ""
license = "Apache-2.0"
authors = ["Florian Ludwig <f.ludwig@greyrook.com>"]
homepage = "https://github.com/FlorianLudwig/cosy"

[tool.cosy]
public = true

[tool.poetry.dependencies]
python = "^3.7"
tomlkit = "^0.7.0"
PyYAML = "^5.4.1"
distro = "^1.5.0"
typer = "^0.3.2"


[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
black = "^20.8b1"
mypy = "^0.812"
pylint = "^2.7.4"

[tool.poetry.scripts]
cosy = 'cosy.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
